rename: nickName -> nickname으로 변경
