+++ frontend/src/domains/pickeat/constants/nickname.ts
